Refactor shared list/tuple functions to objsequence.rs
